+++ c/src/xkms/impl/XKMSRegisterResultImpl.cpp
@@ -233,7 +233,7 @@ XKMSRSAKeyPair * XKMSRegisterResultImpl::getRSAKeyPair(const char * passPhrase) 
 
 	// Yep!  Load the key
 	unsigned char kbuf[XSEC_MAX_HASH_SIZE];
-	unsigned int len = CalculateXKMSKEK((unsigned char *) passPhrase, strlen(passPhrase), kbuf, XSEC_MAX_HASH_SIZE);
+	unsigned int len = CalculateXKMSKEK((unsigned char *) passPhrase, (int) strlen(passPhrase), kbuf, XSEC_MAX_HASH_SIZE);
 
 	XSECProvider prov;
 	XENCCipher * cipher = prov.newCipher(m_msg.mp_env->getParentDocument());
@@ -329,7 +329,7 @@ XENCEncryptedData * XKMSRegisterResultImpl::setRSAKeyPair(const char * passPhras
 	}
 
 	unsigned char kbuf[XSEC_MAX_HASH_SIZE];
-	unsigned int len = CalculateXKMSKEK((unsigned char *) passPhrase, strlen(passPhrase), kbuf, XSEC_MAX_HASH_SIZE);
+	unsigned int len = CalculateXKMSKEK((unsigned char *) passPhrase, (int) strlen(passPhrase), kbuf, XSEC_MAX_HASH_SIZE);
 
 	XSECCryptoKey * sk = handler->createKeyForURI(
 					uri,
